Update PhysicsTeacher.java
making a small change to the function of the parent method, in the inherited PhysicsTeacher class, which in turn also changed the output to the child QuantamPhysicsTeacher class. However, this override did NOT make any changes to the parent class, the Teacher.java. This only overrode the inherited class where the method was made, and it's child.